+++ src/util.rs
@@ -4,6 +4,7 @@ use crate::svd::{Access, Cluster, Register, RegisterCluster, RegisterInfo};
 use inflections::Inflect;
 use proc_macro2::{Ident, Literal, Span, TokenStream};
 use quote::{quote, ToTokens};
+use std::path::PathBuf;
 
 use anyhow::{anyhow, bail, Result};
 
@@ -13,7 +14,7 @@ pub const BITS_PER_BYTE: u32 = 8;
 /// that are not valid in Rust ident
 const BLACKLIST_CHARS: &[char] = &['(', ')', '[', ']', '/', ' ', '-'];
 
-#[derive(Clone, Copy, PartialEq, Default, Debug)]
+#[derive(Clone, PartialEq, Debug)]
 pub struct Config {
     pub target: Target,
     pub nightly: bool,
@@ -21,6 +22,21 @@ pub struct Config {
     pub make_mod: bool,
     pub const_generic: bool,
     pub ignore_groups: bool,
+    pub output_dir: PathBuf,
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        Self {
+            target: Target::default(),
+            nightly: false,
+            generic_mod: false,
+            make_mod: false,
+            const_generic: false,
+            ignore_groups: false,
+            output_dir: PathBuf::from("."),
+        }
+    }
 }
 
 #[allow(clippy::upper_case_acronyms)]